requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
prompt-toolkit==3.0.19
traitlets==5.11.2
simber==0.2.6
torch==2.1.0
transformers==4.29.2